backend/python-worker/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
httpx==0.25.2
redis==5.0.0
pymongo==4.6.0
supabase==2.4.0
python-dotenv==1.0.0
asyncpg==0.29.0
pgvector==0.2.5
tenacity==8.2.3
tiktoken==0.5.2
